Rename partition to partitioning
