+++ jgrapht-core/src/main/java/org/jgrapht/alg/interfaces/PartitioningAlgorithm.java
@@ -1,0 +1,152 @@
+/*
+ * (C) Copyright 2018-2018, by Alexandru Valeanu and Contributors.
+ *
+ * JGraphT : a free Java graph-theory library
+ *
+ * This program and the accompanying materials are dual-licensed under
+ * either
+ *
+ * (a) the terms of the GNU Lesser General Public License version 2.1
+ * as published by the Free Software Foundation, or (at your option) any
+ * later version.
+ *
+ * or (per the licensee's choosing)
+ *
+ * (b) the terms of the Eclipse Public License v1.0 as published by
+ * the Eclipse Foundation.
+ */
+package org.jgrapht.alg.interfaces;
+
+import java.io.Serializable;
+import java.util.*;
+import java.util.stream.Collectors;
+
+/**
+ * Algorithm to compute a vertex partitioning of a graph.
+ *
+ * @param <V> vertex the graph vertex type
+ *
+ * @author Alexandru Valeanu
+ */
+public interface PartitioningAlgorithm<V> {
+
+    /**
+     * Computes a vertex partitioning.
+     *
+     * @return a vertex partitioning
+     */
+    Partitioning<V> getPartitioning();
+
+    /**
+     * Check if the given vertex partitioning is valid.
+     *
+     * @param partitioning the input vertex partitioning
+     * @return true if the input partitioning is valid, false otherwise
+     */
+    boolean isValidPartitioning(Partitioning<V> partitioning);
+
+    /**
+     * A <a href="https://en.wikipedia.org/wiki/Graph_partition">graph partitioning</a>.
+     *
+     * @param <V> the vertex type
+     */
+    interface Partitioning<V> extends Iterable<Set<V>>{
+
+        /**
+         * Get the number of partitions.
+         *
+         * @return the number of partitions
+         */
+        int getNumberPartitions();
+
+        /**
+         * Get the index-th partition (0-based).
+         *
+         * @param index index of the partition to return
+         * @return the index-th partition
+         * @throws IndexOutOfBoundsException if the index is out of range
+         *         (<tt>index &lt; 0 || index &gt;= getNumberPartitions()</tt>)
+         */
+        Set<V> getPartition(int index);
+
+        /**
+         * Get the partitions. This method returns a partitioning of
+         * the vertices in the graph into disjoint partitions.
+         *
+         * @return a list of partitions
+         */
+        default List<Set<V>> getPartitions(){
+            final int n = getNumberPartitions();
+            List<Set<V>> partitions = new ArrayList<>(n);
+
+            for (int i = 0; i < n; i++) {
+                partitions.add(getPartition(i));
+            }
+
+            return partitions;
+        }
+    }
+
+    /**
+     * Default implementation of a vertex partition
+     *
+     * @param <V> the vertex type
+     */
+    class PartitioningImpl<V> implements Partitioning<V>, Serializable {
+
+        private static final long serialVersionUID = 3702471090706836080L;
+
+        /* Partitioning classes */
+        private final List<Set<V>> classes;
+
+        /**
+         * Construct a new vertex partitioning.
+         *
+         * @param classes the partition classes
+         * @throws NullPointerException if {@code classes} is {@code null}
+         */
+        public PartitioningImpl(List<Set<V>> classes) {
+            this.classes = Collections.unmodifiableList(Objects.requireNonNull(classes)
+                    .stream()
+                    .map(Collections::unmodifiableSet)
+                    .collect(Collectors.toList()));
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public int getNumberPartitions() {
+            return classes.size();
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public Set<V> getPartition(int index) {
+            if (index < 0 || index >= classes.size()){
+                throw new IndexOutOfBoundsException(index + " is not valid");
+            }
+
+            return classes.get(index);
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public String toString()
+        {
+            return "Partition [number-of-partitions=" + getNumberPartitions() + ", partitions=" + classes + "]";
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public Iterator<Set<V>> iterator() {
+            return classes.iterator();
+        }
+    }
+}
