+++ jgrapht-core/src/test/java/org/jgrapht/GraphTestsTest.java
@@ -17,35 +17,19 @@
  */
 package org.jgrapht;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Random;
-import java.util.Set;
-
 import org.jgrapht.generate.CompleteGraphGenerator;
-import org.jgrapht.generate.GnpRandomBipartiteGraphGenerator;
 import org.jgrapht.generate.NamedGraphGenerator;
 import org.jgrapht.generate.StarGraphGenerator;
-import org.jgrapht.graph.DefaultDirectedGraph;
-import org.jgrapht.graph.DefaultEdge;
-import org.jgrapht.graph.DefaultUndirectedGraph;
-import org.jgrapht.graph.DefaultUndirectedWeightedGraph;
-import org.jgrapht.graph.DefaultWeightedEdge;
-import org.jgrapht.graph.DirectedPseudograph;
-import org.jgrapht.graph.Pseudograph;
-import org.jgrapht.graph.SimpleDirectedGraph;
-import org.jgrapht.graph.SimpleGraph;
+import org.jgrapht.graph.*;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.util.Arrays;
+import java.util.Random;
+
 import static junit.framework.TestCase.fail;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 /**
  * Test class GraphTests.
@@ -407,115 +391,6 @@ public class GraphTestsTest
                         g.addEdge(i, j);
             }
             assertTrue(GraphTests.isSplit(g));
-        }
-    }
-
-    @Test public void testBipartite1()
-    {
-        Graph<Integer, DefaultEdge> g = new Pseudograph<>(DefaultEdge.class);
-        assertTrue(GraphTests.isBipartite(g));
-        g.addVertex(1);
-        assertTrue(GraphTests.isBipartite(g));
-        g.addVertex(2);
-        assertTrue(GraphTests.isBipartite(g));
-        g.addEdge(1, 2);
-        assertTrue(GraphTests.isBipartite(g));
-        g.addVertex(3);
-        assertTrue(GraphTests.isBipartite(g));
-        g.addEdge(2, 3);
-        assertTrue(GraphTests.isBipartite(g));
-        g.addEdge(3, 1);
-        assertFalse(GraphTests.isBipartite(g));
-    }
-
-    @Test public void testBipartite2()
-    {
-        Graph<Integer, DefaultEdge> g = new Pseudograph<>(DefaultEdge.class);
-
-        for (int i = 0; i < 100; i++) {
-            g.addVertex(i);
-            if (i > 0) {
-                g.addEdge(i, i - 1);
-            }
-        }
-        g.addEdge(99, 0);
-        assertTrue(GraphTests.isBipartite(g));
-    }
-
-    @Test public void testBipartite3()
-    {
-        Graph<Integer, DefaultEdge> g = new Pseudograph<>(DefaultEdge.class);
-
-        for (int i = 0; i < 101; i++) {
-            g.addVertex(i);
-            if (i > 0) {
-                g.addEdge(i, i - 1);
-            }
-        }
-        g.addEdge(100, 0);
-        assertFalse(GraphTests.isBipartite(g));
-    }
-
-    @Test public void testBipartite4()
-    {
-        Graph<Integer, DefaultEdge> g = new DirectedPseudograph<>(DefaultEdge.class);
-
-        for (int i = 0; i < 101; i++) {
-            g.addVertex(i);
-            if (i > 0) {
-                g.addEdge(i, i - 1);
-            }
-        }
-        g.addEdge(100, 0);
-        assertFalse(GraphTests.isBipartite(g));
-    }
-
-    @Test public void testRandomBipartite()
-    {
-        GnpRandomBipartiteGraphGenerator<Integer, DefaultEdge> generator =
-            new GnpRandomBipartiteGraphGenerator<>(10, 10, 0.8);
-        for (int i = 0; i < 100; i++) {
-            Graph<Integer, DefaultEdge> g = GraphTestsUtils.createPseudograph();
-            generator.generateGraph(g);
-            assertTrue(GraphTests.isBipartite(g));
-        }
-    }
-
-    @Test public void testIsBipartitePartition()
-    {
-        List<Graph<Integer, DefaultEdge>> gList = new ArrayList<>();
-        gList.add(new Pseudograph<>(DefaultEdge.class));
-        gList.add(new DirectedPseudograph<>(DefaultEdge.class));
-
-        for (Graph<Integer, DefaultEdge> g : gList) {
-            Set<Integer> a = new HashSet<>();
-            Graphs.addAllVertices(g, Arrays.asList(1, 2, 3, 4));
-            a.addAll(Arrays.asList(1, 2));
-            Set<Integer> b = new HashSet<>();
-            b.addAll(Arrays.asList(3, 4));
-            assertTrue(GraphTests.isBipartitePartition(g, a, b));
-            g.addEdge(1, 3);
-            g.addEdge(1, 4);
-            g.addEdge(1, 3);
-            g.addEdge(2, 3);
-            g.addEdge(2, 4);
-            g.addEdge(4, 1);
-            g.addEdge(3, 1);
-            assertTrue(GraphTests.isBipartitePartition(g, a, b));
-            a.remove(1);
-            assertFalse(GraphTests.isBipartitePartition(g, a, b));
-            a.add(1);
-            assertTrue(GraphTests.isBipartitePartition(g, a, b));
-            DefaultEdge e11 = g.addEdge(1, 1);
-            assertFalse(GraphTests.isBipartitePartition(g, a, b));
-            g.removeEdge(e11);
-            assertTrue(GraphTests.isBipartitePartition(g, a, b));
-            DefaultEdge e44 = g.addEdge(4, 4);
-            assertFalse(GraphTests.isBipartitePartition(g, a, b));
-            g.removeEdge(e44);
-            assertTrue(GraphTests.isBipartitePartition(g, a, b));
-            g.addEdge(4, 3);
-            assertFalse(GraphTests.isBipartitePartition(g, a, b));
         }
     }
